+++ examples/bs4.R
@@ -43,6 +43,12 @@ biom_filter <- cohortBuilder::filter(
   "multi_discrete", name = "Biomarkers", variables = c("biom1", "biom2"), dataset = "patients",
   values = NA
 )
+query_filter <- cohortBuilder::filter(
+  "query", id = "iris_query", dataset = "patients", variables = c("group", "gender"),
+  value = queryBuilder::queryGroup(
+    queryBuilder::queryRule("group", "equal", "B")
+  )
+)
 
 pin_controlbar <- function(controlbar, ...) {
   controlbar[[2]] <- controlbar[[2]] %>%
@@ -76,7 +82,8 @@ shiny::runApp(list(
         age_filter,
         treatment_filter,
         visit_filter,
-        biom_filter
+        biom_filter,
+        query_filter
       )
     )
     # or below if initialized without default data source